### 2024-05-23 #6676 @warrickball
Hi @DavoGrant, I've just picked this up and will start looking for suitable reviewers. If you already know of some potential reviewers who don't represent a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), feel free to post them here or email me.
