### 2024-05-28 #6676 @warrickball
Thanks @nenasedk for agreeing to review. As discussed, we've noted your publication with the ExoTiC-LD authors in [Tsai et al. (2023)](https://ui.adsabs.harvard.edu/abs/2023Natur.617..483T) as a potential conflict of interest but it's from a large collaboration and didn't work closely together.
