### 2024-05-31 #6816 @editorialbot
[REVIEW]: ExoTiC-LD: thirty seconds to stellar limb-darkening coefficients
**Submitting author:** @DavoGrant (<a href="http://orcid.org/0000-0001-5878-618X">David Grant</a>)
**Repository:** https://github.com/Exo-TiC/ExoTiC-LD
**Branch with paper.md** (empty if default branch):
**Version:** v3.2.1
