### 2024-06-03 2baab8a
add power-2 LD law
