### 2024-06-03 #6816 @LorenzoMugnai
Hello @DavoGrant,

First of all, congratulations on your work. The package is very well put together, and I see it is already widely used by the community.

I was reviewing your documentation and have a few comments and suggestions for improvements, in line with the JOSS guidelines:

- **Community Guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?

  I didn't see a section addressing this in the documentation. You could add a paragraph on the homepage to guide readers on how to contribute or contact you if they encounter any issues with the code.

- **Functionality Documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

  Your API section has excellent and complete docstrings, and I see you aimed to make the API section intuitive rather than just a list of functions. I respect that, but it could benefit from some text to guide users in understanding where to look for specific information or how your code is structured. Currently, I see a list of functions separated into different pages. For example, you could use a single page as a "commented" index for all your functions, which would be easier to navigate.

- **Installation Instructions:** Is there a clearly stated list of dependencies? Ideally, these should be handled with an automated package management solution.

  You clearly state how to install the code from PyPI, but since it is available on GitHub, what if someone wants to install it from the source to edit it? You should add a section on how to install from the source.

Thank you for considering these suggestions, and great work once again!
