### 2024-06-04 897d194
Find missing doi if exist.
