+++ joss/paper.bib
@@ -203,6 +203,7 @@ archivePrefix = {arXiv},
   journal={Astronomy \& Astrophysics},
   volume={553},
   year={2013},
+  doi = {10.1051/0004-6361/201219058},
   publisher={EDP Sciences}
 }
 
@@ -222,6 +223,7 @@ archivePrefix = {arXiv},
   volume={573},
   pages={A90},
   year={2015},
+  doi = {10.1051/0004-6361/201423804},
   publisher={EDP Sciences}
 }
 
@@ -233,6 +235,7 @@ archivePrefix = {arXiv},
   number={3},
   pages={39},
   year={2023},
+  doi = {10.3847/2515-5172/acc180},
   publisher={The American Astronomical Society}
 }
 
@@ -243,6 +246,7 @@ archivePrefix = {arXiv},
   volume={666},
   pages={A60},
   year={2022},
+  doi = {10.1051/0004-6361/202243722},
   publisher={EDP Sciences}
 }
 
@@ -277,7 +281,8 @@ archivePrefix = {arXiv},
   volume={5},
   number={46},
   pages={1834},
-  year={2020}
+  year={2020},
+  doi = {10.21105/joss.01834}
 }
 
 @software{Bourque:2021,
