### 2024-06-05 1682bc8
Add contributing guidelines to docs.
