+++ docs/source/views/contributing.rst
@@ -1,6 +1,6 @@
-# Contribution guidelines
+Contributing guidelines
+=======================
 
-## General guidelines
 Hello and thank you for contributing to our project! Go ahead open a new issue or new pull request (PR) for bugs,
 feedback, or new features you would like to see by following these guidelines.
 Leave a comment anywhere and we will be happy to assist. New contributions and contributors are very welcome!
@@ -10,8 +10,9 @@ We recommend that you **fork** the repository and submit your requests according
 2) Open a separate issue or pull request for each suggestion or bug report.
 3) Pull requests, issues and commits should have a useful title.
 4) Give every PR a description and open it against **dev**, the development branch of the repository.
-5) Every PR that adds or changes a feature should also add or change the corresponding **test**, if applicable, as well
-as **documentation**.
+5) Every PR that adds or changes a feature should also add or change the corresponding **test**, if applicable, as well as **documentation**.
 
-## Code of conduct
-This package follows the Contributor Covenant [Code of Conduct](CODE_OF_CONDUCT.md) to provide a welcoming community to everybody.
+Code of conduct
+---------------
+This package follows the Contributor Covenant `Code of Conduct <https://github.com/Exo-TiC/ExoTiC-LD/blob/main/CODE_OF_CONDUCT.md>`_
+to provide a welcoming community to everybody.
