### 2024-06-05 b05c3b9
Added ld law eqns to docstrings.
