### 2024-06-06 #6816 @LorenzoMugnai
> Hi @LorenzoMugnai -- thanks for taking the time to review our repository. We really appreciate the work that goes into reviewing.
>
> We have addressed each of your initial comments, and I think the documentation is looking much improved thanks to your suggestions. Here is what we have done:
>
> * Community guidelines: We have added a “contributing” page to the documentation and signposted to this from the homepage. We have also updated the contributing.md file in the package root.
> * Functionality documentation: We have added all the class methods and submodule functions to the API index page. In this page we have added text to help orientate the user. Additionally, we have improved the docstrings for the limb-darkening laws, now showing the latex-typed functional forms in the docs.
> * Installation instructions: We have added a subsection on installing from source to the installation page in the documentation.

Hi @DavoGrant  I just saw your updated documentation: it looks so much better to me. Great work!
