### 2024-06-10 #6816 @nenasedk
Hi @DavoGrant

Great writeup, and ExoTIC-LD was super easy to install and get running. I have a few small recommendations for your consideration:

1. Statement of need: what does ExoTic provide that other limb darkening codes (Espinoza limb-darkening, LDTk) do not?

2. What is specific about this version that justifies the JOSS publication? Were there recently major updates, or is the JOSS paper meant to cover ExoTic in general? If so, maybe a bit more detail about the features would be appropriate (e.g. probabilistic computations).

3. API documentation: there are functions for 3 parameter, 4 parameter limb darkening coefficients. Maybe add a line to the API documentation stating that the coefficients can be found in exotic_ld.ls_laws.whatever_the_law_is.

4. It would be helpful if the tutorials explained why the user should choose one parameterisation or another.

5. Bibliography/DOI links should be included for the Husser, Kostogryz N, Kostogryz NM, Magic and Morello papers.

6. Line 20 and throughout: either use limb-darkening or limb darkening (hyphenation)

7. Line 22: Minor quibble - Then retrieve the wavelength-dependent... should be rephrased as Then the wavelength-dependent sensitivity must be retrieved... or Then one retrieves... or something to that effect, right now the sentence lacks a subject.

8. phoenix models are usually stylised as PHOENIX models. (all caps)
