+++ docs/source/views/api/api.rst
@@ -11,8 +11,11 @@ Primary Interface
 
 The primary class in ExoTiC-LD is StellarLimbDarkening. On instantiation, this
 class handles the stellar models. Next, there are methods for computing the
-limb-darkening coefficients utilising the aforementioned stellar models. A list of
-all the methods are given below.
+limb-darkening coefficients utilising the aforementioned stellar models. The
+coefficients returned by each method can be found by clicking into the list
+of methods. Additionally, the functional forms and coefficients for each
+limb-darkening law can be found in the :doc:`ld_laws <./ld_laws>` subpackage
+below.
 
 .. toctree::
     :maxdepth: 2
@@ -23,8 +26,8 @@ Subpackages
 -----------
 
 The available limb-darkening laws are listed here. These functions are utilised by
-the corresponding StellarLimbDarkening methods shown above. Functional forms of each
-law can be seen by clicking into the functions.
+the corresponding StellarLimbDarkening methods shown above. Functional forms and
+coefficients for each law can be inspected by clicking into the functions.
 
 .. toctree::
     :maxdepth: 2
