### 2024-06-12 #6816 @DavoGrant
Hi @nenasedk -- many thanks for your suggestions and your quick turn around on this. Let me go through each point in turn (same numbering as above):

1. Our aim with ExoTiC-LD was to make a complete and frictionless tool for limb-darkening computation. Ease of use was our top priority, and this seems to be working based on usership in the community. Beyond this, the code provides access to a more extensive collection of stellar grids than most, and under the hood we have developed a tree structure for selecting stellar models within the 3D parameters space. I feel the statement of need already speaks to this, but let us know if you had any specific additions in mind.
2. This JOSS submission is for ExoTiC-LD in general. Publication was prompted by reaching feature completeness and an uptick in usership.
3. We have added additional signposts in the main API docs page to the ld_laws subpackage. We have added a few words explaining that this is where to inspect the functional forms and coefficients for each limb-darkening law.
4. We have added some text to the first tutorial explaining that choosing a parameterisation is very much application dependent. We signpost to some good literature that goes into these choices in more detail.
5. Done. See commit 897d194.
6. Hyphenation is (I believe) used when it is a compound modifier; e.g., limb-darkening coefficient.
7. Done.
8. Done.
