### 2024-06-25 #6816 @LorenzoMugnai
Hello, sorry for my late reply.
I was testing the tutorials and I found an error raised for an apparently missing file. Following the JOSS directions, I raised an [Issue](https://github.com/Exo-TiC/ExoTiC-LD/issues/49) ticket in the repository  with (hopefully) all the information to reproduce the issue.

Other than that, the software and the manuscript look very good to me.
Great job!
