### 2024-07-02 Exo-TiC/ExoTiC-LD#49 @DavoGrant
Thanks @LorenzoMugnai. Looks like this is a simple capitalisation bug. "JWST_NIRSpec_prism" should be "JWST_NIRSpec_Prism", I have just fixed this in the tutorials. For backwards compatibility, I have also added the lowercase version to the data server to prevent any earlier users running into the same error.
