### 2024-07-02 #6816 @DavoGrant
@LorenzoMugnai, bug now fixed in (see https://github.com/Exo-TiC/ExoTiC-LD/pull/50).
