### 2024-07-02 #6816 @LorenzoMugnai
> @LorenzoMugnai, bug now fixed in (see [Exo-TiC/ExoTiC-LD#50](https://github.com/Exo-TiC/ExoTiC-LD/pull/50)).

Briliant! Thank you!
I have no other comments. I completed my checklist.
Thanks
