### 2024-07-04 #6816 @LorenzoMugnai
Yes, I am happy to recommend the package for publication
