### 2024-07-10 #6816 @warrickball
Hi @DavoGrant, I've had a look at the paper myself and I think you should expand on what ExoTiC-LD does that other codes don't. I appreciate you've mentioned four other codes but it raises the question: why do we need another one? I don't doubt there's a good reason—compactness or ease of use alone are enough—but it should be included in the paper.

Also, the [ASCL entry for JKTLD](https://ascl.net/1511.016) suggests citing [Southworth (2008)](https://ui.adsabs.harvard.edu/abs/2008MNRAS.386.1644S) as the appropriate reference. This seems to be more widely used, with ~334 citations vs ~18 for the ASCL record. (You're welcome to cite both.)
