### 2024-07-22 d914e69
Make comparison to prev software more explicit.
