+++ joss/paper.md
@@ -46,16 +46,23 @@ frustrating amount of “data admin”. In brief, one starts with a search throu
 grids of stellar models to find a good match with the science target in 
 metallicity, effective temperature, and surface gravity. Then, one must retrieve 
 the wavelength-dependent sensitivity of the employed instrument, process all 
-these data into the correct form, and then finally compute the limb-darkening 
+these data into a cohesive form, and then finally compute the limb-darkening 
 coefficients.
 
-In `ExoTiC-LD` we have done all of the heavy lifting for the user, making the 
-process as fast and frictionless as possible. A user simply has to `pip install` 
+Previous software has made calculating limb-darkening coefficients available 
+to the community [e.g., @Southworth:2015; @Parviainen:2015; @Morello:2020; @Bourque:2021], 
+albeit with varying degrees of installation complexity and access to stellar and 
+instrument data. In `ExoTiC-LD` we have done all of the heavy lifting for the user, 
+making the process as fast and frictionless as possible. A user simply has to `pip install` 
 the code and the relevant data will be automatically downloaded at runtime and 
-the limb-darkening coefficients computed. In particular, all the stellar and 
-instrument data has been pre-processed and homogenised. Additionally, the 
+the limb-darkening coefficients computed. In particular, a wide selection of stellar 
+and instrument data has been pre-processed and homogenised. Additionally, the 
 stellar model grids have been stored as tree structures, enabling an efficient 
-search for good matches and helpful warnings to the user.
+search for good matches and helpful warnings to the user. Currently, the stellar 
+models supported are PHOENIX [@Husser:2013], kurucz [@Kurucz:1993], 
+stagger [@Magic:2015], and MPS-ATLAS [@Kostogryz:2022; @Kostogryz:2023]. 
+There are also options to provide custom data if the user has their own stellar 
+models or instrument data.
 
 `ExoTiC-LD` thus far has predominantly been utilised in the study of exoplanet 
 atmospheres, helping to facilitate the study of Jupiter-like
@@ -64,18 +71,12 @@ and Earth-like exoplanets [e.g., @Moran:2023; @Kirk:2024]. It has
 also been incorporated into the popular open-source JWST data reduction and 
 analysis pipeline, called `Eureka!` [@Bell:2022].
 
-Currently, the stellar models supported are PHOENIX [@Husser:2013], kurucz 
-[@Kurucz:1993], stagger [@Magic:2015], and MPS-ATLAS [@Kostogryz:2022; @Kostogryz:2023]. 
-There are also options to provide custom data if the user has their own stellar 
-models or instrument data. Other available codes include those referenced in 
-@Southworth:2015, @Parviainen:2015, @Morello:2020, and @Bourque:2021.
-
 # Acknowledgements
 
 We acknowledge contributions from Natasha Batalha, Matthew Hill, and Iva Laginja, 
 as well as testing by Taylor Bell, Lili Alderson, Daniel Valentine, Charlotte 
-Fairman, and Katy Chubb. D.G. and H.R.W was funded by UK Research and Innovation 
-(UKRI) under the UK government’s Horizon Europe funding guarantee as part of an 
-ERC Starter Grant [grant number EP/Y006313/1].
+Fairman, Katy Chubb, and Nikole Lewis. D.G. and H.R.W were funded by UK Research 
+and Innovation (UKRI) under the UK government’s Horizon Europe funding guarantee 
+as part of an ERC Starter Grant [grant number EP/Y006313/1].
 
 # References
