### 2024-07-22 #6816 @DavoGrant
@warrickball, just made those updates you have suggested.

- All the cited codes include the same core functionality to compute limb-darkening coefficients. As mentioned above, our code really aims at lowering the barrier for entry through the way it handles the data, the user interface, and the installation procedure. We have reordered and updated the paper text to convey this message more explicitly, and I hope now this message is clear for the reader.
- Southworth reference updated to the 2008 citation.

Thanks, let us know the next steps.
