### 2024-07-29 #6816 @warrickball
Hi @DavoGrant, I'm happy with these changes and that the @nenasedk and @LorenzoMugnai have recommended the paper for publication. There's just a bit more housekeeping to do as noted in the **Additional Author Tasks After Review is Complete** above, which mostly boils down to making a tagged release of the code and archiving that version in a Zenodo repository. (I'll check off the editor tasks as we proceed.)

Let me know if you need any extra guidance.
