### 2024-08-05 2c119f2
Update README.rst
changed zenodo link to v3.2.1 JOSS publication version
