### 2024-08-05 faa26bc
Update README.rst
updated Zenodo reference
