### 2024-08-06 #6816 @DavoGrant
@warrickball this is great news.

@hrwakeford has made a Zenodo release, this is v3.2.1, and has the following DOI https://doi.org/10.5281/zenodo.13224465. We have checked the author, affiliations, and licenses too.
