### 2024-08-07 #6816 @DavoGrant
> While I check the Zenodo archive, I just noticed that some of the references which you presumably got from NASA ADS aren't showing the journal names because of the shorthands like `\apj` and `\aap`. Could you replace these with the journal names (e.g. Astrophysical Journal, Astronomy & Astrophysics)? These don't warrant a new release (though you can make one if you like).

Done.
