### 2024-08-07 #6816 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
